api/pyproject.toml
[tool.poetry]
name = "gccsite-api"
version = "0.1.0"
description = "Backend of Girls Can Code! website"
authors = ["Association Prologin <info@prologin.org>"]
license = "AGPL-3.0"

repository = "https://gitlab.com/prologin/gcc/site"

packages = [
    { include = "gccsite" }
]

[tool.poetry.dependencies]
python = "^3.10"

django = "^3.2.7"
django-cors-headers = "^3.8.0"
djangorestframework = "^3.12.4"
django-filter = "^21.1"
social-auth-backend-prologin = "^0.1.3"
django-storages = "^1.11.1"
boto3 = "^1.18.46"
drf-yasg = "^1.20.0"
jsonschema = "^4.13.0"
djangorestframework-simplejwt = "^5.2.0"
psycopg2 = "^2.8.4"

[tool.poetry.dev-dependencies]
pre-commit = "^2"
black = "^22"
flake8 = "^4"
isort = "^5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
preview = true
exclude = ["migrations"]

[tool.isort]
profile = "black"
line_length = "79"
